requirements.txt
botcity-framework-core>=1.2.1,<2.0
botcity-framework-web>=0.9.0,<1.0
botcity-email-plugin>=0.1.5
botcity-maestro-sdk>=0.3.3,<1.0
imap-tools>=1.7.2
# Uncomment to use botcity-utils
# botcity-utils>=0.1.2

# Uncomment to use botcity-crawler-plugin
# botcity-crawler-plugin

# Uncomment to use Pandas
# pandas
